Modulariza funciones y organiza estructura del proyecto
- Se modularizaron las funciones creadas, moviendo las características al
    archivo main y sus respectivos tests a la carpeta designada.
- Se implementó la lógica para consultar e impactar en la base de datos.
- Se agregó el método "get_available" a la clase GameRepository.

 Cambios a ser confirmados:
	borrados:        listar_partidas_disponibles.py
	modificados:     main.py
	modificados:     model/game.py
	modificados:     repositories/game.py
	borrados:        test_listar_partidas_disponibles.py
	modificados:     tests/test_sanity.py
