+++ model/game.py
@@ -68,6 +68,8 @@ class Game(Base):
             raise GameFull
         self.players.append(player)
 
+    def count_players(self) -> int:
+        return len(self.players)
 
 class GameFull(BaseException):
     pass
