+++ repositories/game.py
@@ -1,6 +1,6 @@
 from typing import Optional, List
 from repositories.general import Repository
-from model import Game
+from model import Game, Player
 
 
 class GameRepository(Repository):
@@ -18,3 +18,7 @@ class GameRepository(Repository):
 
     def get_many(self, count: int) -> List[Game]:
         return self.db.query(Game).limit(count).all()
+
+    def get_available(self, count: int) -> List[Game]:
+        games = self.db.query(Game).filter(Game.started == False).all()
+        return [game for game in games if len(game.players) < game.max_players][:count]
